Split out to framework where possible
