+++ CSharpSourceCode/CampaignMechanics/Careers/TORCareerPerkCampaignBehavior.cs
@@ -9,10 +9,10 @@ using TaleWorlds.CampaignSystem.Roster;
 using TaleWorlds.Core;
 using TaleWorlds.Library;
 using TaleWorlds.Localization;
-using TOR_Core.CampaignMechanics.Crafting;
 using TOR_Core.CharacterDevelopment;
 using TOR_Core.CharacterDevelopment.CareerSystem;
 using TOR_Core.Extensions;
+using TOR_Core.Framework;
 using TOR_Core.Items;
 using TOR_Core.Utilities;
 using static TOR_Core.Utilities.TORConstants;
