+++ CSharpSourceCode/CampaignMechanics/Crafting/EnchantingVM.cs
@@ -8,6 +8,7 @@ using TaleWorlds.Library;
 using TaleWorlds.LinQuick;
 using TOR_Core.CampaignMechanics.Crafting.Models;
 using TOR_Core.Extensions;
+using TOR_Core.Framework;
 using TOR_Core.Items;
 using TOR_Core.Utilities;
