+++ CSharpSourceCode/CampaignMechanics/Crafting/TORArtisanDistrictCampaignBehavior.cs
@@ -10,8 +10,8 @@ using TaleWorlds.Core;
 using TaleWorlds.Localization;
 using TaleWorlds.ObjectSystem;
 using TaleWorlds.SaveSystem;
-using TOR_Core.CampaignMechanics.TORCustomSettlement;
 using TOR_Core.Extensions;
+using TOR_Core.Framework;
 using TOR_Core.Items;
 using TOR_Core.Utilities;
 using static TaleWorlds.CampaignSystem.CampaignBehaviors.CraftingCampaignBehavior;
