+++ CSharpSourceCode/CampaignMechanics/TORCustomSettlement/GreenskinBrawlBehavior.cs
@@ -12,6 +12,7 @@ using TaleWorlds.Core;
 using TaleWorlds.Localization;
 using TaleWorlds.ObjectSystem;
 using TOR_Core.Extensions;
+using TOR_Core.Framework;
 using TOR_Core.Missions;
 using TOR_Core.Utilities;
